Refactors into notifications into smaller components, fixes reload bug
