+++ resources/js/Layouts/Partials/Notifications.tsx
@@ -1,10 +1,9 @@
-import { Link, router, usePage, WhenVisible } from '@inertiajs/react';
+import { router, usePage, WhenVisible } from '@inertiajs/react';
 import LoadingIndicator from '@/Components/LoadingIndicator';
 import { useEffect, useState } from 'react';
 import { Notification } from '@/types/notification/Notification';
-import timeSince from '@/Filters/timeSince';
 import NotificationActions from '@/Layouts/Partials/NotificationComponents/NotificationActions';
-import { TrashIcon } from '@heroicons/react/24/outline';
+import NotificationList from '@/Layouts/Partials/NotificationComponents/NotificationList';
 
 export default function Notifications() {
     const { notifications, notifications_paginator } = usePage().props;
@@ -16,16 +15,9 @@ export default function Notifications() {
         setUnreadNotifications(notifications?.filter(({ read_at }) => read_at === null) ?? []);
     }, [notifications]);
 
-    const deleteNotification = (id: string) => {
-        try {
-            router.delete(route('notifications.destroy', { notification: id }), {
-                onSuccess: () =>
-                    router.reload({ only: ['notifications, notifications_paginator'] }),
-            });
-        } catch (error) {
-            console.error(error);
-        }
-    };
+    useEffect(() => {
+        router.reload({ only: [], data: { notifications: '' } });
+    }, []);
 
     return (
         <>
@@ -38,84 +30,11 @@ export default function Notifications() {
                 {notifications && notifications?.length > 0 ? (
                     <>
                         {unreadNotifications.length > 0 && (
-                            <div className="relative">
-                                <div className="sticky top-0 z-10 border-y border-b-gray-200 border-t-gray-100 bg-gray-50 px-3 py-1.5 text-sm/6 font-semibold text-gray-900">
-                                    <h3>New</h3>
-                                </div>
-                                <ul role="list" className="divide-y divide-gray-100">
-                                    {unreadNotifications.map((notification) => (
-                                        <Link
-                                            key={notification.id}
-                                            as="li"
-                                            href={route(notification.data.route, {
-                                                ...notification.data.params,
-                                                notification: notification.id,
-                                            })}
-                                            className="flex py-5 w-full  hover:bg-gray-200 hover:cursor-pointer"
-                                        >
-                                            <div className="flex items-center justify-between w-full px-4">
-                                                <div className="w-full px-2">
-                                                    <p className="text-sm/6 font-semibold text-gray-900">
-                                                        {notification.data.message}
-                                                    </p>
-                                                    <p className="mt-1 truncate text-xs/5 text-gray-500">
-                                                        {timeSince(notification.created_at)}
-                                                    </p>
-                                                </div>
-                                                <button
-                                                    onClick={(e) => {
-                                                        e.stopPropagation();
-                                                        deleteNotification(notification.id);
-                                                    }}
-                                                    className="p-2"
-                                                >
-                                                    <TrashIcon className="size-6 text-upei-red-500 hover:text-upei-red-700" />
-                                                </button>
-                                            </div>
-                                        </Link>
-                                    ))}
-                                </ul>
-                            </div>
+                            <NotificationList notifications={unreadNotifications} title="New" />
                         )}
 
                         {readNotifications.length > 0 && (
-                            <div className="relative">
-                                <div className="sticky top-0 z-10 border-y border-b-gray-200 border-t-gray-100 bg-gray-50 px-3 py-1.5 text-sm/6 font-semibold text-gray-900">
-                                    <h3>Old</h3>
-                                </div>
-                                <ul role="list" className="divide-y divide-gray-100">
-                                    {readNotifications.map((notification) => (
-                                        <Link
-                                            key={notification.id}
-                                            as="li"
-                                            href={route(
-                                                notification.data.route,
-                                                notification.data.params
-                                            )}
-                                            className="flex justify-center py-5  hover:bg-gray-200 hover:cursor-pointer"
-                                        >
-                                            <div className="flex items-center justify-between w-full px-4">
-                                                <div className="w-full px-2">
-                                                    <p className="text-sm/6 font-semibold text-gray-900">
-                                                        {notification.data.message}
-                                                    </p>
-                                                    <p className="mt-1 truncate text-xs/5 text-gray-500">
-                                                        {timeSince(notification.created_at)}
-                                                    </p>
-                                                </div>
-                                                <button
-                                                    onClick={(e) => {
-                                                        e.stopPropagation();
-                                                        deleteNotification(notification.id);
-                                                    }}
-                                                >
-                                                    <TrashIcon className="size-6 text-upei-red-500 hover:text-upei-red-700" />
-                                                </button>
-                                            </div>
-                                        </Link>
-                                    ))}
-                                </ul>
-                            </div>
+                            <NotificationList notifications={readNotifications} title="Seen" />
                         )}
                         {notifications_paginator &&
                             notifications_paginator.current_page <
